Destructable Objects Script pt.2
Destroyed Objects can now be replaced by a prefab and explode, with the exploding pieces fading away after a short period.

New issue: damage state rendering is broken, so the original rendering will be disabled leaving nothing... but the object can still be destroyed. Will continue work on this tomorrow

I'm so cold and hungry