Destructible Objects Script pt.3
Now there is a scriptable damage state that will spawn when reaching a certain HP threshold that then deactivates when the object is completely destroyed.

There is a millisecond lag when the damage state prefabs are activated, however. Later down the line, the instantiate will be replaced with simple activateObject commands to save performance